fix: Remove HTMLMotionProps import and unused discord variable
- Use standard React HTML props instead of HTMLMotionProps
- Remove unused discord variable to fix eslint errors
- Fixes SyntaxError preventing app from loading
