+++ src/components/ui/Button.tsx
@@ -1,8 +1,8 @@
-import { HTMLMotionProps, motion } from 'framer-motion';
+import { motion } from 'framer-motion';
 import { forwardRef } from 'react';
 import { cn } from '../../utils/cn';
 
-interface ButtonProps extends HTMLMotionProps<'button'> {
+interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
   variant?: 'primary' | 'secondary' | 'dice';
   size?: 'sm' | 'md' | 'lg';
 }
